Add tests and coverage report
