+++ .gitignore
@@ -7,6 +7,7 @@ package-lock.json
 .vercel
 /.svelte-kit
 /build
+/coverage
 
 # OS
 .DS_Store
